+++ packages/cli/src/workflows/workflows.controller.ts
@@ -14,8 +14,8 @@ import { validateEntity } from '@/GenericHelpers';
 import { ExternalHooks } from '@/ExternalHooks';
 import type { ListQuery, WorkflowRequest } from '@/requests';
 import { isBelowOnboardingThreshold } from '@/WorkflowHelpers';
-import { EEWorkflowController } from './workflows.controller.ee';
 import { WorkflowService } from './workflow.service';
+import { isSharingEnabled, rightDiff } from '@/UserManagement/UserManagementHelper';
 import { Container } from 'typedi';
 import { InternalHooks } from '@/InternalHooks';
 import { RoleService } from '@/services/role.service';
@@ -30,9 +30,13 @@ import { NotFoundError } from '@/errors/response-errors/not-found.error';
 import { InternalServerError } from '@/errors/response-errors/internal-server.error';
 import { NamingService } from '@/services/naming.service';
 import { TagRepository } from '@/databases/repositories/tag.repository';
+import { EnterpriseWorkflowService } from './workflow.service.ee';
+import { WorkflowRepository } from '@/databases/repositories/workflow.repository';
+import type { RoleNames } from '@/databases/entities/Role';
+import { UnauthorizedError } from '@/errors/response-errors/unauthorized.error';
+import { CredentialsService } from '../credentials/credentials.service';
 
 export const workflowsController = express.Router();
-workflowsController.use('/', EEWorkflowController);
 
 /**
  * POST /workflows
@@ -61,6 +65,24 @@ workflowsController.post(
 		await WorkflowHelpers.replaceInvalidCredentials(newWorkflow);
 
 		WorkflowHelpers.addNodeIds(newWorkflow);
+
+		if (isSharingEnabled()) {
+			// This is a new workflow, so we simply check if the user has access to
+			// all used workflows
+
+			const allCredentials = await CredentialsService.getMany(req.user);
+
+			try {
+				Container.get(EnterpriseWorkflowService).validateCredentialPermissionsToUser(
+					newWorkflow,
+					allCredentials,
+				);
+			} catch (error) {
+				throw new BadRequestError(
+					'The workflow you are trying to save contains credentials that are not shared with you',
+				);
+			}
+		}
 
 		let savedWorkflow: undefined | WorkflowEntity;
 
@@ -112,7 +134,8 @@ workflowsController.get(
 	listQueryMiddleware,
 	async (req: ListQuery.Request, res: express.Response) => {
 		try {
-			const sharedWorkflowIds = await WorkflowHelpers.getSharedWorkflowIds(req.user, ['owner']);
+			const roles: RoleNames[] = isSharingEnabled() ? [] : ['owner'];
+			const sharedWorkflowIds = await WorkflowHelpers.getSharedWorkflowIds(req.user, roles);
 
 			const { workflows: data, count } = await Container.get(WorkflowService).getMany(
 				sharedWorkflowIds,
@@ -195,6 +218,37 @@ workflowsController.get(
 	ResponseHelper.send(async (req: WorkflowRequest.Get) => {
 		const { id: workflowId } = req.params;
 
+		if (isSharingEnabled()) {
+			const relations = ['shared', 'shared.user', 'shared.role'];
+			if (!config.getEnv('workflowTagsDisabled')) {
+				relations.push('tags');
+			}
+
+			const workflow = await Container.get(WorkflowRepository).get(
+				{ id: workflowId },
+				{ relations },
+			);
+
+			if (!workflow) {
+				throw new NotFoundError(`Workflow with ID "${workflowId}" does not exist`);
+			}
+
+			const userSharing = workflow.shared?.find((shared) => shared.user.id === req.user.id);
+			if (!userSharing && !req.user.hasGlobalScope('workflow:read')) {
+				throw new UnauthorizedError(
+					'You do not have permission to access this workflow. Ask the owner to share it with you',
+				);
+			}
+
+			const enterpriseWorkflowService = Container.get(EnterpriseWorkflowService);
+
+			enterpriseWorkflowService.addOwnerAndSharings(workflow);
+			await enterpriseWorkflowService.addCredentialsToWorkflow(workflow, req.user);
+			return workflow;
+		}
+
+		// sharing disabled
+
 		const extraRelations = config.getEnv('workflowTagsDisabled') ? [] : ['workflow.tags'];
 
 		const shared = await Container.get(SharedWorkflowRepository).findSharing(
@@ -226,18 +280,27 @@ workflowsController.patch(
 	'/:id(\\w+)',
 	ResponseHelper.send(async (req: WorkflowRequest.Update) => {
 		const { id: workflowId } = req.params;
+		const forceSave = req.query.forceSave === 'true';
 
-		const updateData = new WorkflowEntity();
+		let updateData = new WorkflowEntity();
 		const { tags, ...rest } = req.body;
 		Object.assign(updateData, rest);
+
+		if (isSharingEnabled()) {
+			updateData = await Container.get(EnterpriseWorkflowService).preventTampering(
+				updateData,
+				workflowId,
+				req.user,
+			);
+		}
 
 		const updatedWorkflow = await Container.get(WorkflowService).update(
 			req.user,
 			updateData,
 			workflowId,
 			tags,
-			true,
-			['owner'],
+			isSharingEnabled() ? forceSave : true,
+			isSharingEnabled() ? undefined : ['owner'],
 		);
 
 		return updatedWorkflow;
@@ -274,10 +337,101 @@ workflowsController.delete(
 workflowsController.post(
 	'/run',
 	ResponseHelper.send(async (req: WorkflowRequest.ManualRun): Promise<IExecutionPushResponse> => {
+		if (isSharingEnabled()) {
+			const workflow = Container.get(WorkflowRepository).create(req.body.workflowData);
+
+			if (req.body.workflowData.id !== undefined) {
+				const safeWorkflow = await Container.get(EnterpriseWorkflowService).preventTampering(
+					workflow,
+					workflow.id,
+					req.user,
+				);
+				req.body.workflowData.nodes = safeWorkflow.nodes;
+			}
+		}
+
 		return Container.get(WorkflowService).runManually(
 			req.body,
 			req.user,
 			GenericHelpers.getSessionId(req),
+		);
+	}),
+);
+
+/**
+ * (EE) PUT /workflows/:id/share
+ *
+ * Grant or remove users' access to a workflow.
+ */
+workflowsController.put(
+	'/:workflowId/share',
+	ResponseHelper.send(async (req: WorkflowRequest.Share) => {
+		if (!isSharingEnabled()) throw new NotFoundError('Route not found');
+
+		const { workflowId } = req.params;
+		const { shareWithIds } = req.body;
+
+		if (
+			!Array.isArray(shareWithIds) ||
+			!shareWithIds.every((userId) => typeof userId === 'string')
+		) {
+			throw new BadRequestError('Bad request');
+		}
+
+		const isOwnedRes = await Container.get(EnterpriseWorkflowService).isOwned(req.user, workflowId);
+		const { ownsWorkflow } = isOwnedRes;
+		let { workflow } = isOwnedRes;
+
+		if (!ownsWorkflow || !workflow) {
+			workflow = undefined;
+			// Allow owners/admins to share
+			if (req.user.hasGlobalScope('workflow:share')) {
+				const sharedRes = await Container.get(WorkflowService).getSharing(req.user, workflowId, {
+					allowGlobalScope: true,
+					globalScope: 'workflow:share',
+				});
+				workflow = sharedRes?.workflow;
+			}
+			if (!workflow) {
+				throw new UnauthorizedError('Forbidden');
+			}
+		}
+
+		const ownerIds = (
+			await Container.get(WorkflowRepository).getSharings(
+				Db.getConnection().createEntityManager(),
+				workflowId,
+				['shared', 'shared.role'],
+			)
+		)
+			.filter((e) => e.role.name === 'owner')
+			.map((e) => e.userId);
+
+		let newShareeIds: string[] = [];
+		await Db.transaction(async (trx) => {
+			// remove all sharings that are not supposed to exist anymore
+			await Container.get(WorkflowRepository).pruneSharings(trx, workflowId, [
+				...ownerIds,
+				...shareWithIds,
+			]);
+
+			const sharings = await Container.get(WorkflowRepository).getSharings(trx, workflowId);
+
+			// extract the new sharings that need to be added
+			newShareeIds = rightDiff(
+				[sharings, (sharing) => sharing.userId],
+				[shareWithIds, (shareeId) => shareeId],
+			);
+
+			if (newShareeIds.length) {
+				await Container.get(EnterpriseWorkflowService).share(trx, workflow!, newShareeIds);
+			}
+		});
+
+		void Container.get(InternalHooks).onWorkflowSharingUpdate(
+			workflowId,
+			req.user.id,
+			shareWithIds,
 		);
 	}),
 );
